simplify LSP setup
